Microservices/FixSuggest/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.12
PyYAML==6.0.2
